Explicitly add two hooks to `next` entry point
- Rename `keyboard-navigation` -> `use-arrow-key-navigation` to follow
  naming conventions
